Add pyproject and move torchchat.py into torchchat
