+++ pyproject.toml
@@ -1,0 +1,11 @@
+[project]
+name = "torchchat"
+description = "A native-PyTorch library for LLM Inference"
+readme = "README.md"
+version = "0.1.0"
+
+[project.scripts]
+torchchat = "torchchat.torchchat:main"
+
+[tool.setuptools]
+packages = ["torchchat"]
