+++ torchchat/cli/builder.py
@@ -17,7 +17,7 @@ from torch.distributed.device_mesh import DeviceMesh
 import torch._dynamo.config
 import torch._inductor.config
 
-from config.model_config import resolve_model_config
+from torchchat.model_params.model_config import resolve_model_config
 from distributed import init_distributed, ParallelDims, parallelize_llama
 from torchchat.utils.quantize import quantize_model
 from torchchat.utils.measure_time import measure_time
