+++ torchchat/model_params/model_config.py
@@ -61,7 +61,7 @@ def load_model_configs() -> Dict[str, ModelConfig]:
     model_configs = {}
 
     with open(
-        Path(__file__).parent.parent / "config" / "data" / "models.json", "r"
+        Path(__file__).parent/ "models.json", "r"
     ) as f:
         model_config_dict = json.load(f)
