+++ torchchat/usages/__init__.py
@@ -1,3 +1,3 @@
-from torchchat.usages import browser, openai_api, server
+from torchchat.usages import browser, openai_api, server, eval
 
-__all__ = [browser, openai_api, server]
+__all__ = [browser, openai_api, server, eval]
