+++ torchchat/usages/eval.py
@@ -19,7 +19,7 @@ from torchchat.cli.builder import (
 from torchchat.model import Transformer
 from torchchat.utils.build_utils import set_precision
 from torchchat.cli.cli import add_arguments_for_verb, arg_init
-from utils.measure_time import measure_time
+from torchchat.utils.measure_time import measure_time
 
 torch._dynamo.config.automatic_dynamic_shapes = True
 torch._inductor.config.triton.unique_kernel_names = True
